{{cookiecutter.git_repo_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.git-repo-name}}"
version = "0.1.0"
description = ""
authors = ["Ryan Whitten <rwhitten577@gmail.com>"]
packages = [
     { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
boto3 = "^1.16.3"
python-dotenv = "^0.14.0"
uvicorn = "^0.12.2"
alembic = "^1.4.3"
SQLAlchemy = "^1.3.20"
python-jose = {version = "^3.2.0", extras = ["cryptography"]}
requests = "^2.24.0"
psycopg2 = "^2.8.6"
pydantic = {version = "^1.6.1", extras = ["email"]}
slackclient = "^2.9.3"
dateparser = "^0.7.6"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
coverage = "^5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
